refactor(st-09051): modularize multi-agent agent
