+++ packages/patterns/src/multi-agent/agent.ts
@@ -1,56 +1,19 @@
 /**
  * Multi-Agent System Factory
  *
- * This module provides the main factory function for creating multi-agent systems.
+ * This module provides the public facade for creating multi-agent systems.
  *
  * @module patterns/multi-agent/agent
  */
 
-import { StateGraph, END, CompiledStateGraph } from '@langchain/langgraph';
-import type { RunnableConfig } from '@langchain/core/runnables';
-import { createPatternLogger } from '../shared/deduplication.js';
-import { MultiAgentState } from './state.js';
-import type { MultiAgentStateType } from './state.js';
-import type { MultiAgentSystemConfig, MultiAgentRouter, WorkerConfig } from './types.js';
-import { createSupervisorNode, createWorkerNode, createAggregatorNode } from './nodes.js';
+import { createCompiledMultiAgentSystem } from './agent-graph.js';
+import { registerWorkerCapabilities } from './agent-runtime.js';
+import type { MultiAgentSystemWithRegistry, RegisterWorkerInput } from './agent-types.js';
 import type { WorkerCapabilities } from './schemas.js';
+import type { MultiAgentSystemConfig } from './types.js';
 
-const logger = createPatternLogger('agentforge:patterns:multi-agent:system');
-
-type ToolLike = {
-  metadata?: { name?: string };
-  name?: string;
-};
-
-/**
- * Extract tool name from either AgentForge Tool or LangChain tool
- *
- * AgentForge Tools have: tool.metadata.name
- * LangChain tools have: tool.name
- *
- * @param tool - Tool instance (AgentForge or LangChain)
- * @returns Tool name or 'unknown' if not found
- */
-function getToolName(tool: unknown): string {
-  if (!tool || typeof tool !== 'object') {
-    return 'unknown';
-  }
-
-  const candidate = tool as ToolLike;
-
-  // AgentForge Tool: has metadata.name
-  if (typeof candidate.metadata?.name === 'string') {
-    return candidate.metadata.name;
-  }
-
-  // LangChain tool: has name directly
-  if (typeof candidate.name === 'string') {
-    return candidate.name;
-  }
-
-  // Fallback
-  return 'unknown';
-}
+export { MultiAgentSystemBuilder } from './agent-builder.js';
+export type { MultiAgentSystemWithRegistry, RegisterWorkerInput } from './agent-types.js';
 
 /**
  * Create a multi-agent coordination system
@@ -137,285 +100,7 @@ function getToolName(tool: unknown): string {
  * ```
  */
 export function createMultiAgentSystem(config: MultiAgentSystemConfig): MultiAgentSystemWithRegistry {
-  const {
-    supervisor,
-    workers,
-    aggregator,
-    maxIterations = 10,
-    verbose = false,
-    checkpointer,
-  } = config;
-
-  // Note: Empty workers array is allowed - workers can be registered later with registerWorkers()
-
-  // Create the graph
-  const workflow = new StateGraph(MultiAgentState);
-
-  const supervisorConfig = { ...supervisor, maxIterations, verbose };
-
-  // Add supervisor node
-  const supervisorNode = createSupervisorNode(supervisorConfig);
-  workflow.addNode('supervisor', supervisorNode);
-
-  // Add worker nodes and collect capabilities
-  const workerIds: string[] = [];
-  const workerCapabilities: Record<string, WorkerCapabilities> = {};
-
-  for (const workerConfig of workers) {
-    const workerNode = createWorkerNode({
-      ...workerConfig,
-      verbose,
-    });
-    workflow.addNode(workerConfig.id, workerNode);
-    workerIds.push(workerConfig.id);
-
-    // Store worker capabilities for state initialization
-    workerCapabilities[workerConfig.id] = workerConfig.capabilities;
-  }
-
-  // Add aggregator node
-  const aggregatorNode = createAggregatorNode({
-    ...aggregator,
-    verbose,
-  });
-  workflow.addNode('aggregator', aggregatorNode);
-
-  // Define routing logic
-  const supervisorRouter: MultiAgentRouter = (state: MultiAgentStateType) => {
-    logger.debug('Supervisor router executing', {
-      status: state.status,
-      ...(state.currentAgent ? { currentAgent: state.currentAgent } : {}),
-      iteration: state.iteration
-    });
-
-    // Check for completion or failure
-    if (state.status === 'completed' || state.status === 'failed') {
-      logger.info('Supervisor router: ending workflow', { status: state.status });
-      return END;
-    }
-
-    // Check if we should aggregate
-    if (state.status === 'aggregating') {
-      logger.info('Supervisor router: routing to aggregator');
-      return 'aggregator';
-    }
-
-    // Route to the current agent(s)
-    // Support parallel routing: currentAgent may contain comma-separated agent IDs
-    if (state.currentAgent && state.currentAgent !== 'supervisor') {
-      // Check if this is parallel routing (multiple agents)
-      if (state.currentAgent.includes(',')) {
-        // Return array of agent IDs for parallel execution
-        const agents = state.currentAgent.split(',').map(a => a.trim());
-        logger.info('Supervisor router: parallel routing', {
-          agents,
-          count: agents.length
-        });
-        return agents; // LangGraph supports returning arrays for parallel execution
-      }
-      // Single agent routing
-      logger.info('Supervisor router: single agent routing', {
-        targetAgent: state.currentAgent
-      });
-      return state.currentAgent;
-    }
-
-    // Default: stay at supervisor
-    logger.debug('Supervisor router: staying at supervisor');
-    return 'supervisor';
-  };
-
-  const workerRouter: MultiAgentRouter = (state: MultiAgentStateType) => {
-    logger.debug('Worker router executing', {
-      iteration: state.iteration,
-      completedTasks: state.completedTasks.length
-    });
-
-    // Workers always return to supervisor
-    // The supervisor will check if all work is done and route to aggregator
-    logger.debug('Worker router: returning to supervisor');
-    return 'supervisor';
-  };
-
-  const aggregatorRouter: MultiAgentRouter = (state: MultiAgentStateType) => {
-    logger.info('Aggregator router: ending workflow', {
-      completedTasks: state.completedTasks.length,
-      status: state.status
-    });
-
-    // Aggregator always ends
-    return END;
-  };
-
-  // Set entry point
-  // @ts-expect-error - LangGraph StateGraph generic mismatch with string node names
-  workflow.setEntryPoint('supervisor');
-
-  // Add edges from supervisor
-  // @ts-expect-error - LangGraph StateGraph generic mismatch with string node names
-  workflow.addConditionalEdges('supervisor', supervisorRouter, [
-    'aggregator',
-    END,
-    ...workerIds,
-  ]);
-
-  // Add edges from workers back to supervisor
-  for (const workerId of workerIds) {
-    // @ts-expect-error - LangGraph StateGraph generic mismatch with dynamic node names
-    workflow.addConditionalEdges(workerId, workerRouter, ['supervisor']);
-  }
-
-  // Add edge from aggregator to end
-  // @ts-expect-error - LangGraph StateGraph generic mismatch with string node names
-  workflow.addConditionalEdges('aggregator', aggregatorRouter, [END]);
-
-  // Compile the graph with checkpointer if provided
-  const compiled = workflow.compile(checkpointer ? { checkpointer } : undefined) as unknown as MultiAgentSystemWithRegistry;
-
-  // Wrap the invoke method to inject worker capabilities into the initial state
-  const originalInvoke = compiled.invoke.bind(compiled);
-  compiled.invoke = (async function(input: Partial<MultiAgentStateType>, config?: RunnableConfig) {
-    // Merge worker capabilities with any workers in the input
-    const mergedInput = {
-      ...input,
-      workers: {
-        ...workerCapabilities,
-        ...(input.workers || {}),
-      },
-    };
-
-    return originalInvoke(
-      mergedInput as Parameters<typeof originalInvoke>[0],
-      config as Parameters<typeof originalInvoke>[1]
-    );
-  }) as unknown as typeof compiled.invoke;
-
-  // Wrap the stream method to inject worker capabilities into the initial state
-  const originalStream = compiled.stream.bind(compiled);
-  compiled.stream = (async function(input: Partial<MultiAgentStateType>, config?: RunnableConfig) {
-    // Merge worker capabilities with any workers in the input
-    const mergedInput = {
-      ...input,
-      workers: {
-        ...workerCapabilities,
-        ...(input.workers || {}),
-      },
-    };
-
-    return originalStream(
-      mergedInput as Parameters<typeof originalStream>[0],
-      config as Parameters<typeof originalStream>[1]
-    );
-  }) as unknown as typeof compiled.stream;
-
-  return compiled as MultiAgentSystemWithRegistry;
-}
-
-/**
- * Multi-agent system builder for dynamic worker registration
- *
- * This builder allows you to register workers before compiling the graph.
- * Once compiled, the graph is immutable and workers cannot be added.
- */
-export class MultiAgentSystemBuilder {
-  private config: MultiAgentSystemConfig;
-  private additionalWorkers: WorkerConfig[] = [];
-  private compiled: boolean = false;
-
-  constructor(config: Omit<MultiAgentSystemConfig, 'workers'> & { workers?: WorkerConfig[] }) {
-    this.config = {
-      ...config,
-      workers: config.workers || [],
-    };
-  }
-
-  /**
-   * Register workers with the system builder
-   *
-   * @param workers - Array of worker configurations
-   * @returns this builder for chaining
-   *
-   * @example
-   * ```typescript
-   * const builder = new MultiAgentSystemBuilder({
-   *   supervisor: { llm, strategy: 'skill-based' },
-   *   aggregator: { llm },
-   * });
-   *
-   * builder.registerWorkers([
-   *   {
-   *     name: 'math_worker',
-   *     capabilities: ['math', 'calculations'],
-   *     tools: [calculatorTool],
-   *   },
-   * ]);
-   *
-   * const system = builder.build();
-   * ```
-   */
-  registerWorkers(workers: Array<{
-    name: string;
-    description?: string;
-    capabilities: string[];
-    tools?: WorkerConfig['tools'];
-    systemPrompt?: string;
-    model?: WorkerConfig['model'];
-  }>): this {
-    if (this.compiled) {
-      throw new Error('Cannot register workers after the system has been compiled');
-    }
-
-    // Convert to WorkerConfig format
-    for (const worker of workers) {
-      this.additionalWorkers.push({
-        id: worker.name,
-        capabilities: {
-          skills: worker.capabilities,
-          tools: worker.tools?.map(t => getToolName(t)) || [],
-          available: true,
-          currentWorkload: 0,
-        },
-        model: worker.model || this.config.supervisor.model,
-        tools: worker.tools,
-        systemPrompt: worker.systemPrompt,
-      });
-    }
-    return this;
-  }
-
-  /**
-   * Build and compile the multi-agent system
-   *
-   * @returns Compiled LangGraph workflow
-   */
-  build() {
-    if (this.compiled) {
-      throw new Error('System has already been compiled');
-    }
-
-    // Merge configured workers with registered workers
-    const allWorkers = [...this.config.workers, ...this.additionalWorkers];
-
-    if (allWorkers.length === 0) {
-      throw new Error('At least one worker must be registered before building the system');
-    }
-
-    this.compiled = true;
-
-    return createMultiAgentSystem({
-      ...this.config,
-      workers: allWorkers,
-    });
-  }
-}
-
-/**
- * Extended multi-agent system with worker registration support
- */
-export interface MultiAgentSystemWithRegistry extends CompiledStateGraph<MultiAgentStateType, unknown> {
-  _workerRegistry?: Record<string, WorkerCapabilities>;
-  _originalInvoke?: MultiAgentSystemWithRegistry['invoke'];
-  _originalStream?: MultiAgentSystemWithRegistry['stream'];
+  return createCompiledMultiAgentSystem(config);
 }
 
 /**
@@ -447,76 +132,14 @@ export interface MultiAgentSystemWithRegistry extends CompiledStateGraph<MultiAg
  */
 export function registerWorkers(
   system: MultiAgentSystemWithRegistry,
-  workers: Array<{
-    name: string;
-    description?: string;
-    capabilities: string[];
-    tools?: WorkerConfig['tools'];
-    systemPrompt?: string;
-  }>
+  workers: RegisterWorkerInput[],
 ): void {
   console.warn(
     '[AgentForge] registerWorkers() on a compiled system only updates worker capabilities in state.\n' +
     'It does NOT add worker nodes to the graph. Use MultiAgentSystemBuilder for proper worker registration.\n' +
     'See: https://github.com/TVScoundrel/agentforge/blob/main/packages/patterns/docs/multi-agent-pattern.md'
   );
-
-  // Initialize registry if it doesn't exist
-  if (!system._workerRegistry) {
-    system._workerRegistry = {};
-  }
-
-  // Convert worker configs to capabilities format
-  for (const worker of workers) {
-    system._workerRegistry[worker.name] = {
-      skills: worker.capabilities,
-      tools: worker.tools?.map(t => getToolName(t)) || [],
-      available: true,
-      currentWorkload: 0,
-    };
-  }
-
-  // Wrap the invoke method to inject workers into state (only once)
-  if (!system._originalInvoke) {
-    system._originalInvoke = system.invoke.bind(system);
-
-    system.invoke = (async function(input: Partial<MultiAgentStateType>, config?: RunnableConfig) {
-      // Merge registered workers with any workers in the input
-      const mergedInput = {
-        ...input,
-        workers: {
-          ...(system._workerRegistry || {}),
-          ...(input.workers || {}),
-        },
-      };
-
-      return system._originalInvoke!(
-        mergedInput as Parameters<NonNullable<typeof system._originalInvoke>>[0],
-        config as Parameters<NonNullable<typeof system._originalInvoke>>[1]
-      );
-    }) as unknown as typeof system.invoke;
-  }
-
-  // Wrap the stream method to inject workers into state (only once)
-  if (!system._originalStream) {
-    system._originalStream = system.stream.bind(system);
-
-    system.stream = (async function(input: Partial<MultiAgentStateType>, config?: RunnableConfig) {
-      // Merge registered workers with any workers in the input
-      const mergedInput = {
-        ...input,
-        workers: {
-          ...(system._workerRegistry || {}),
-          ...(input.workers || {}),
-        },
-      };
-
-      return system._originalStream!(
-        mergedInput as Parameters<NonNullable<typeof system._originalStream>>[0],
-        config as Parameters<NonNullable<typeof system._originalStream>>[1]
-      );
-    }) as unknown as typeof system.stream;
-  }
+  registerWorkerCapabilities(system, workers);
 }
 
 /**
